update python project use setup tools, and upgrade github action version
