--- Pipfile
@@ -1,11 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-docker = "*"
-python-dotenv = "*"
-dnspython = "*"
