+++ docker_nsupdate_ddns/lib/nsupdate.py
@@ -1,19 +1,20 @@
+import logging
+
 import dns.update
 import dns.query
 import dns.tsigkeyring
 import ipaddress
-from datetime import datetime
 
 config = {}
 
+LOG = logging.getLogger(__name__)
+
 
 def add_records(records):
-    keyring = dns.tsigkeyring.from_text(
-        {config['TSIG_NAME']: config['TSIG_KEY']})
+    keyring = dns.tsigkeyring.from_text({config['TSIG_NAME']: config['TSIG_KEY']})
 
     for hostname, ip in records.items():
-        print(datetime.now().isoformat(), end=" ")
-        print("Adding record for " + hostname + "(" + ip + ")")
+        LOG.info(f"Adding record for {hostname}({ip})")
 
         rrtype = "A"
         address = ipaddress.ip_address(ip)
@@ -28,12 +29,10 @@ def add_records(records):
 
 
 def delete_records(records):
-    keyring = dns.tsigkeyring.from_text(
-        {config['TSIG_NAME']: config['TSIG_KEY']})
+    keyring = dns.tsigkeyring.from_text({config['TSIG_NAME']: config['TSIG_KEY']})
 
     for hostname, ip in records.items():
-        print(datetime.now().isoformat(), end=" ")
-        print("Deleting record for " + hostname + "(" + ip + ")")
+        LOG.info(f"Deleting record for {hostname}({ip})")
 
         update = dns.update.Update(config['DOMAIN'], keyring=keyring)
         update.delete(hostname)
